refactor: rename spawnSync.js to spawn-sync.js
I named this file wrong originally, before I knew that the convention in
this repo was to use kebab-case file names for this kind of thing.
