+++ packages/liferay-npm-scripts/__tests__/scripts/theme.js
@@ -30,8 +30,8 @@ describe('scripts/theme.js', () => {
 
 	beforeEach(() => {
 		cwd = process.cwd();
-		jest.mock('../../src/utils/spawnSync');
-		spawnSync = require('../../src/utils/spawnSync');
+		jest.mock('../../src/utils/spawn-sync');
+		spawnSync = require('../../src/utils/spawn-sync');
 	});
 
 	afterEach(() => {
