+++ packages/liferay-npm-scripts/src/scripts/build.js
@@ -13,7 +13,7 @@ const getMergedConfig = require('../utils/get-merged-config');
 const {moveToTemp, removeFromTemp} = require('../utils/move-to-temp');
 const setEnv = require('../utils/set-env');
 const {buildSoy, cleanSoy, soyExists} = require('../utils/soy');
-const spawnSync = require('../utils/spawnSync');
+const spawnSync = require('../utils/spawn-sync');
 const validateConfig = require('../utils/validateConfig');
 const withBabelConfig = require('../utils/with-babel-config');
 const webpack = require('./webpack');
