+++ packages/liferay-npm-scripts/src/scripts/webpack.js
@@ -5,7 +5,7 @@
  */
 
 const fs = require('fs');
-const spawnSync = require('../utils/spawnSync');
+const spawnSync = require('../utils/spawn-sync');
 const withBabelConfig = require('../utils/with-babel-config');
 
 const WATCH_CONFIG_FILE = 'webpack.config.dev.js';
